+++ src/pages/account/edit-username.tsx
@@ -1,7 +1,7 @@
 import { H4, H5, Text } from "@/modules/common/components/Typography";
 import { useUser } from "@/modules/common/hooks/useUserStore";
+import BackButton from "@/modules/user/account/components/BackButton";
 import styled from "@emotion/styled";
-import { CaretLeft } from "@phosphor-icons/react";
 import { Button, Input } from "antd";
 import Head from "next/head";
 import { useRouter } from "next/router";
@@ -23,10 +23,7 @@ const EditUsername = () => {
       </Head>
       <ScreenContainer>
         <Container>
-          <BackButtonContainer onClick={() => onClickBackButton()}>
-            <CaretLeft size={32} />
-            <Text>กลับสู่หน้าโพรไฟล์</Text>
-          </BackButtonContainer>
+          <BackButton text={"กลับสู่หน้าโพรไฟล์"} />
           <ProfileContainer>
             <H4>แก้ไขชื่อผู้ใช้</H4>
             <Text type="secondary">ป้อนชื่อผู้ใช้ใหม่และรหัสผ่านของคุณ</Text>
@@ -99,18 +96,6 @@ const Container = styled.div`
       ${(props) => props.theme.antd.colorPrimaryBorder} -27.67%,
       #fff 18.01%
     );
-`;
-
-const BackButtonContainer = styled(Button)`
-  display: flex;
-  justify-content: space-between;
-  width: 30%;
-  align-items: center;
-  min-width: 175px;
-  margin-bottom: 12px;
-  border: none;
-  box-shadow: none;
-  background: none;
 `;
 
 const ProfileContainer = styled.div`
